+++ packages/localizations/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 3.13.5
+
+### Patch Changes
+
+- Update fr-FR TOS/privacy policy loclization to include the appropriate links. ([#5511](https://github.com/clerk/javascript/pull/5511)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Add payment source section to `UserProfile` ([#5492](https://github.com/clerk/javascript/pull/5492)) by [@aeliox](https://github.com/aeliox)
+
+- Updated dependencies [[`ec4521b`](https://github.com/clerk/javascript/commit/ec4521b4fe56602f524a0c6d1b09d21aef5d8bd0), [`f30fa75`](https://github.com/clerk/javascript/commit/f30fa750754f19030f932a666d2bdbdf0d86743d), [`9c68678`](https://github.com/clerk/javascript/commit/9c68678e87047e6312b708b775ebfb23a3e22f8a)]:
+  - @clerk/types@4.50.2
+
 ## 3.13.4
 
 ### Patch Changes